+++ test/new_relic/agent/transaction_test.rb
@@ -2026,5 +2026,33 @@ module NewRelic::Agent
 
       assert_nil txn.add_span_link(stub_link)
     end
+
+    def test_span_events_returns_empty_array_when_no_events_added
+      in_transaction do |txn|
+        assert_equal NewRelic::EMPTY_ARRAY, txn.span_events
+      end
+    end
+
+    def test_span_events_returns_empty_array_without_initial_segment
+      txn = Transaction.new(:web, {})
+
+      assert_equal NewRelic::EMPTY_ARRAY, txn.span_events
+    end
+
+    def test_add_span_event_stores_event_on_initial_segment
+      in_transaction do |txn|
+        txn.add_span_event('my_event', attributes: {'k' => 'v'})
+
+        assert_equal 1, txn.span_events.length
+        assert_equal 'my_event', txn.span_events.first[:name]
+        assert_equal txn.initial_segment.span_events, txn.span_events
+      end
+    end
+
+    def test_add_span_event_is_noop_without_initial_segment
+      txn = Transaction.new(:web, {})
+
+      assert_nil txn.add_span_event('ghost_event')
+    end
   end
 end